Update file path for CV in About component
The file path for the CV in the About component has been modified. The previous file 'CV-Frontend-developer.pdf' was replaced by a new file 'CV Frontend-developer.pdf', eliminating the hyphen in the name. This change was necessary as the file name was updated in order to observe proper file naming conventions.